Fix Windows conversion and deprecated API warnings
Make floating-point conversions explicit while preserving existing calculations. Select secure CRT, wide Winsock, and descriptor APIs on Windows without changing the Linux and Tizen paths.

+++ dali/internal/vector-image/common/vector-image-renderer-impl.cpp
@@ -288,8 +288,8 @@ bool VectorImageRenderer::Load(const Vector<uint8_t>& data)
     return false;
   }
 
-  mDefaultWidth  = mParsedImage->width;
-  mDefaultHeight = mParsedImage->height;
+  mDefaultWidth  = static_cast<uint32_t>(mParsedImage->width);
+  mDefaultHeight = static_cast<uint32_t>(mParsedImage->height);
 #endif
 
   DALI_LOG_INFO(gVectorImageLogFilter, Debug::Verbose, "Load success! DefaultSize [%u x %u] [%p]\n", mDefaultWidth, mDefaultHeight, this);
